+++ rust/lance/src/dataset/write/update.rs
@@ -302,7 +302,7 @@ impl UpdateJob {
                     }
                 }
             })
-            .buffer_unordered(self.dataset.object_store.io_parallelism() as usize);
+            .buffer_unordered(self.dataset.object_store.io_parallelism());
 
         while let Some(res) = stream.next().await.transpose()? {
             match res {
